split integration test
