+++ integration/kill_test.go
@@ -1,0 +1,23 @@
+package integration
+
+import (
+	"fmt"
+
+	. "gopkg.in/check.v1"
+)
+
+func (s *RunSuite) TestKill(c *C) {
+	p := s.ProjectFromText(c, "up", SimpleTemplate)
+
+	name := fmt.Sprintf("%s_%s_1", p, "hello")
+
+	cn := s.GetContainerByName(c, name)
+	c.Assert(cn, NotNil)
+	c.Assert(cn.State.Running, Equals, true)
+
+	s.FromText(c, p, "kill", SimpleTemplate)
+
+	cn = s.GetContainerByName(c, name)
+	c.Assert(cn, NotNil)
+	c.Assert(cn.State.Running, Equals, false)
+}
