+++ integration/pull_test.go
@@ -1,0 +1,15 @@
+package integration
+
+import (
+	. "gopkg.in/check.v1"
+)
+
+func (s *RunSuite) TestPull(c *C) {
+	//TODO: This doesn't test much
+	s.ProjectFromText(c, "pull", `
+        hello:
+          image: tianon/true
+          stdin_open: true
+          tty: true
+        `)
+}
